Move modeling to additional/bonus materials
Modeling or writing their own array iteration methods is no longer
an objective.  Effectively using the methods is.  All modeling of the
methods has been moved to the bottom of lesson as additional.  The first
additional lab is to write `count` which if there is time should be done
with the developers but to keep it under 3pts it is listed as additional