Simplify examples to align with objectives
The demos, code alongs and labs now fit the objective for developers
to use array iteration methods.  These examples are much simpler and
do not pull from a CSV which added a layer likely more valuable for
bonus material than understanding the methods themselves.

`lib` starter code now has some basic arrays to start with instead of
requiring and parsing them from a CSV file which added a layer outside
of the scope of the objectives.